FIO.proto ref_block_prefix type changed to uint64 (#879)
* ref_block_prefix type changed to uint64
* remove toInt()
* Add tests for chainParams range checking.
* Move chainParam trunctation to Transaction class, extend to block number as well.
* Minor cleanup.

+++ src/FIO/TransactionBuilder.h
@@ -26,7 +26,7 @@ class ChainParams {
 public:
     Data chainId;
     uint64_t headBlockNumber;
-    uint32_t refBlockPrefix;
+    uint64_t refBlockPrefix;
 };
 
 class TransactionBuilder {
